Refactoring & structure cleanup
