+++ src/App.jsx
@@ -1,7 +1,7 @@
 import './App.css';
 // import { BrowserRouter as Router, Route, Routes, Link } from 'react-router-dom';
 import { createBrowserRouter, RouterProvider } from "react-router-dom";
-import { FilterPage } from "./pages/Home/FilterPage"
+import { FilterPage } from "./pages/Filter/FilterPage"
 import { HomePage } from "./pages/Home/HomePage"
 import { SignupPage } from "./pages/Home/SignupPage";
 import { FavouritesPage } from './pages/Favourites/FavouritesPage';
@@ -10,9 +10,9 @@ import BookPage from './pages/Book/BookPage';
 
 import { SearchPage } from './pages/Search/SearchPage';
 
-import CartPage from './pages/Cart/CartPage';
+import BasketPage from './pages/Basket/BasketPage';
 import OrderSuccessPage from './pages/Checkout/SuccessPage';
-import OrderCancelPage from './pages/Checkout/cancel';
+import OrderCancelPage from './pages/Checkout/CancelPage';
 
 
 
@@ -44,7 +44,7 @@ const router = createBrowserRouter([
   },
   {
     path: "/basket",
-    element: <CartPage />
+    element: <BasketPage />
   },
   {
     path: "/favourites",
