+++ src/components/Structure/Footer.css
@@ -1,4 +1,4 @@
-/* src/Components/Home/Footer.css */
+/* src/Components/Structure/Footer.css */
 
 .footer {
     background-color: #A1CCD1;
