+++ src/pages/Book/BookPage.jsx
@@ -2,8 +2,8 @@
 
 import React, { useEffect, useState } from 'react';
 import { useParams, Link } from 'react-router-dom';
-import Navbar from '../../components/Home/NavBar.jsx';
-import Footer from "../../components/Home/Footer.jsx";
+import Navbar from '../../components/Structure/NavBar.jsx';
+import Footer from "../../components/Structure/Footer.jsx";
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faHeart, faShoppingBasket, faStar } from '@fortawesome/free-solid-svg-icons';
 import './BookPage.css';
