+++ src/services/order.jsx
@@ -1,5 +1,6 @@
 // orderService.js
-const backendUrl = "https://booksy-backend.onrender.com";
+//const backendUrl = "https://booksy-backend.onrender.com";
+const backendUrl = "http://127.0.0.1:3000"
 
 export const createOrderRecord = async (basketDetails) => {
     try {
