+++ src/services/stripe.jsx
@@ -1,4 +1,5 @@
-const backendUrl = "https://booksy-backend.onrender.com"; // Your backend URL
+//const backendUrl = "https://booksy-backend.onrender.com"; // Your backend URL
+const backendUrl  = "http://127.0.0.1:3000"
 
 import {loadStripe} from '@stripe/stripe-js';
